# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Harmony ITSM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List agents (staff who work tickets)
      parameters:
        - in: query
          name: departmentId
          schema: { type: integer }
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Agent" }
    post:
      operationId: createAgent
      tags: [agents]
      summary: Add an agent (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAgentInput" }
      responses:
        "201":
          description: Created agent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
  /agents/{id}:
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update an agent (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAgentInput" }
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Delete an agent (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List hardware assets
      parameters:
        - in: query
          name: departmentId
          schema: { type: integer }
        - in: query
          name: status
          schema:
            type: string
            enum: [in_use, in_storage, retired, repair]
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200":
          description: Assets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Asset" }
    post:
      operationId: createAsset
      tags: [assets]
      summary: Add an asset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAssetInput" }
      responses:
        "201":
          description: Created asset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
    patch:
      operationId: updateAsset
      tags: [assets]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAssetInput" }
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
    delete:
      operationId: deleteAsset
      tags: [assets]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    Role:
      type: string
      enum: [admin, agent, end_user]
...
    Agent:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        title:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        ticketsAssigned: { type: integer }
      required: [id, name, email, role, ticketsAssigned]
    CreateAgentInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        title:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
      required: [name, email, role]
    UpdateAgentInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        title:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
...
    Asset:
      type: object
      properties:
        id: { type: integer }
        assetTag: { type: string }
        name: { type: string }
        type:
          type: string
          enum: [laptop, desktop, monitor, phone, printer, server, tablet, network, tool, vehicle, other]
        manufacturer:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        location: { type: string }
        site:
          type: string
          enum: [office, jobsite]
        status:
          type: string
          enum: [in_use, in_storage, retired, repair]
        assignedToId:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        purchasedAt:
          type: ["string", "null"]
          format: date-time
        warrantyEndsAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, assetTag, name, type, location, site, status, createdAt]
    CreateAssetInput:
      type: object
      properties:
        assetTag: { type: string }
        name: { type: string }
        type:
          type: string
          enum: [laptop, desktop, monitor, phone, printer, server, tablet, network, tool, vehicle, other]
        manufacturer:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        location: { type: string }
        site:
          type: string
          enum: [office, jobsite]
        status:
          type: string
          enum: [in_use, in_storage, retired, repair]
        assignedToId:
          type: ["integer", "null"]
        departmentId:
          type: ["integer", "null"]
        purchasedAt:
          type: ["string", "null"]
          format: date-time
        warrantyEndsAt:
          type: ["string", "null"]
          format: date-time
      required: [assetTag, name, type, location, site, status]
    UpdateAssetInput:
      type: object
      properties:
        assetTag: { type: string }
        name: { type: string }
        type:
          type: string
          enum: [laptop, desktop, monitor, phone, printer, server, tablet, network, tool, vehicle, other]
        manufacturer:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        location: { type: string }
        site:
          type: string
          enum: [office, jobsite]
        status:
          type: string
          enum: [in_use, in_storage, retired, repair]
        assignedToId:
          type: ["integer", "null"]
        departmentId:
          type: ["integer", "null"]
        purchasedAt:
          type: ["string", "null"]
          format: date-time
        warrantyEndsAt:
          type: ["string", "null"]
          format: date-time
...