# lib/api-zod/src/generated/types/createAgentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Harmony ITSM API
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from "./role";

export interface CreateAgentInput {
  name: string;
  email: string;
  role: Role;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  departmentId?: number | null;
}