# lib/api-zod/src/generated/types/updateAssetInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Harmony ITSM API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAssetInputSite } from "./updateAssetInputSite";
import type { UpdateAssetInputStatus } from "./updateAssetInputStatus";
import type { UpdateAssetInputType } from "./updateAssetInputType";

export interface UpdateAssetInput {
  assetTag?: string;
  name?: string;
  type?: UpdateAssetInputType;
  /** @nullable */
  manufacturer?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  location?: string;
  site?: UpdateAssetInputSite;
  status?: UpdateAssetInputStatus;
  /** @nullable */
  assignedToId?: number | null;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  purchasedAt?: Date | null;
  /** @nullable */
  warrantyEndsAt?: Date | null;
}